sandbox - extract IPC based file service (client) and improve runtime behaviour
