+++ src/vs/workbench/services/extensionManagement/browser/builtinExtensionsScannerService.ts
@@ -10,7 +10,7 @@ import { IUriIdentityService } from 'vs/workbench/services/uriIdentity/common/ur
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { URI } from 'vs/base/common/uri';
 import { getGalleryExtensionId } from 'vs/platform/extensionManagement/common/extensionManagementUtil';
-import { getUriFromAmdModule } from 'vs/base/common/amd';
+import { FileAccess } from 'vs/base/common/network';
 
 interface IScannedBuiltinExtension {
 	extensionPath: string;
@@ -70,7 +70,7 @@ export class BuiltinExtensionsScannerService implements IBuiltinExtensionsScanne
 			enableBuiltinExtensions = environmentService.remoteAuthority ? false : true;
 		}
 		if (enableBuiltinExtensions) {
-			return getUriFromAmdModule(require, '../../../../../../extensions');
+			return FileAccess.asBrowserUri('../../../../../../extensions', require);
 		}
 		return undefined;
 	}
